Consolidate YarnUtils with Utils
They are both utils for working with yarn, this change drops Utils
